+++ portal-kernel/test/unit/com/liferay/portal/modules/ModulesStructureTest.java
@@ -68,10 +68,6 @@ public class ModulesStructureTest {
 			classLoader,
 			"com/liferay/portal/modules/dependencies" +
 				"/git_repo_gitattributes.tmpl");
-		final String gitRepoGradlePropertiesTemplate = StringUtil.read(
-			classLoader,
-			"com/liferay/portal/modules/dependencies" +
-				"/git_repo_gradle_properties.tmpl");
 		final String gitRepoSettingsGradleTemplate = StringUtil.read(
 			classLoader,
 			"com/liferay/portal/modules/dependencies" +
@@ -103,7 +99,6 @@ public class ModulesStructureTest {
 						_testGitRepoBuildScripts(
 							dirPath, gitRepoBuildGradleTemplate,
 							gitRepoGitAttributesTemplate,
-							gitRepoGradlePropertiesTemplate,
 							gitRepoSettingsGradleTemplate);
 					}
 					else if (Files.exists(dirPath.resolve("app.bnd"))) {
@@ -481,9 +476,7 @@ public class ModulesStructureTest {
 			buildGradleTemplate, "[$BUILDSCRIPT_DEPENDENCIES$]", sb.toString());
 	}
 
-	private String _getGitRepoGradleProperties(
-		Path dirPath, String gradlePropertiesTemplate) {
-
+	private String _getProjectPathPrefix(Path dirPath) {
 		String projectPathPrefix = String.valueOf(
 			_modulesDirPath.relativize(dirPath));
 
@@ -492,8 +485,7 @@ public class ModulesStructureTest {
 				StringUtil.replace(
 					projectPathPrefix, File.separatorChar, CharPool.COLON);
 
-		return gradlePropertiesTemplate.replace(
-			"[$PROJECT_PATH_PREFIX$]", projectPathPrefix);
+		return projectPathPrefix;
 	}
 
 	private boolean _isInGitRepo(Path dirPath) {
@@ -509,10 +501,24 @@ public class ModulesStructureTest {
 	}
 
 	private String _read(Path path) throws IOException {
+		Assert.assertTrue("Missing " + path, Files.exists(path));
+
 		String s = new String(Files.readAllBytes(path), StandardCharsets.UTF_8);
 
 		return StringUtil.replace(
 			s, System.lineSeparator(), StringPool.NEW_LINE);
+	}
+
+	private Properties _readProperties(Path path) throws IOException {
+		Assert.assertTrue("Missing " + path, Files.exists(path));
+
+		Properties properties = new Properties();
+
+		try (InputStream inputStream = Files.newInputStream(path)) {
+			properties.load(inputStream);
+		}
+
+		return properties;
 	}
 
 	private void _testAntPluginIgnoreFiles(Path dirPath) throws IOException {
@@ -532,9 +538,6 @@ public class ModulesStructureTest {
 	private void _testAppBuildScripts(Path dirPath) throws IOException {
 		Path buildGradlePath = dirPath.resolve("build.gradle");
 
-		Assert.assertTrue(
-			"Missing " + buildGradlePath, Files.exists(buildGradlePath));
-
 		String buildGradle = _read(buildGradlePath);
 
 		Assert.assertEquals(
@@ -543,8 +546,6 @@ public class ModulesStructureTest {
 
 	private void _testEquals(Path path, String expected) throws IOException {
 		if (Validator.isNotNull(expected)) {
-			Assert.assertTrue("Missing " + path, Files.exists(path));
-
 			String actual = _read(path);
 
 			Assert.assertEquals("Incorrect " + path, expected, actual);
@@ -586,19 +587,12 @@ public class ModulesStructureTest {
 
 	private void _testGitRepoBuildScripts(
 			Path dirPath, String buildGradleTemplate,
-			String gitAttributesTemplate, String gradlePropertiesTemplate,
-			String settingsGradleTemplate)
+			String gitAttributesTemplate, String settingsGradleTemplate)
 		throws IOException {
 
 		Path buildGradlePath = dirPath.resolve("build.gradle");
 		Path gradlePropertiesPath = dirPath.resolve("gradle.properties");
 		Path settingsGradlePath = dirPath.resolve("settings.gradle");
-
-		boolean buildGradleExists = Files.exists(buildGradlePath);
-		boolean gradlePropertiesExists = Files.exists(gradlePropertiesPath);
-		boolean settingsGradleExists = Files.exists(settingsGradlePath);
-
-		Assert.assertTrue("Missing " + buildGradlePath, buildGradleExists);
 
 		String buildGradle = _read(buildGradlePath);
 
@@ -606,18 +600,13 @@ public class ModulesStructureTest {
 			"Incorrect " + buildGradlePath,
 			_getGitRepoBuildGradle(dirPath, buildGradleTemplate), buildGradle);
 
-		Assert.assertTrue(
-			"Missing " + gradlePropertiesPath, gradlePropertiesExists);
-
-		String gradleProperties = _read(gradlePropertiesPath);
+		Properties gradleProperties = _readProperties(gradlePropertiesPath);
 
 		Assert.assertEquals(
-			"Incorrect " + gradlePropertiesPath,
-			_getGitRepoGradleProperties(dirPath, gradlePropertiesTemplate),
-			gradleProperties);
-
-		Assert.assertTrue(
-			"Missing " + settingsGradlePath, settingsGradleExists);
+			"Incorrect \"" + _PROJECT_PATH_PREFIX + "\" in " +
+				gradlePropertiesPath,
+			_getProjectPathPrefix(dirPath),
+			gradleProperties.getProperty(_PROJECT_PATH_PREFIX));
 
 		String settingsGradle = _read(settingsGradlePath);
 
@@ -643,9 +632,6 @@ public class ModulesStructureTest {
 		throws IOException {
 
 		Path gitIgnorePath = dirPath.resolve(".gitignore");
-
-		Assert.assertTrue(
-			"Missing " + gitIgnorePath, Files.exists(gitIgnorePath));
 
 		String gitIgnore = _read(gitIgnorePath);
 
@@ -683,9 +669,6 @@ public class ModulesStructureTest {
 
 		Path gitIgnorePath = dirPath.resolve(".gitignore");
 
-		Assert.assertTrue(
-			"Missing " + gitIgnorePath, Files.exists(gitIgnorePath));
-
 		String gitIgnore = _read(gitIgnorePath);
 
 		Assert.assertEquals(
@@ -694,6 +677,8 @@ public class ModulesStructureTest {
 
 	private static final String _APP_BUILD_GRADLE =
 		"apply plugin: \"com.liferay.app.defaults.plugin\"";
+
+	private static final String _PROJECT_PATH_PREFIX = "project.path.prefix";
 
 	private static final String _SOURCE_FORMATTER_IGNORE_FILE_NAME =
 		"source_formatter.ignore";