+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -714,6 +714,15 @@ moved to [Liferay Gradle Plugins].
 ### Changed
 - [LPS-69492]: Update the [Liferay Gradle Plugins] dependency to version 3.0.45.
 
+## 2.0.10 - 2016-12-02
+
+### Added
+- [LPS-69501]: Allow portal tool versions to be overridden in a
+`gradle.properties` file contained in any parent directory of the project, for
+example:
+
+		com.liferay.source.formatter.version=1.0.300
+
 [Liferay CDN]: https://cdn.lfrs.sl/repository.liferay.com/nexus/content/groups/public
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
 [Liferay Gradle Plugins App Javadoc Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-app-javadoc-builder
@@ -779,6 +788,7 @@ moved to [Liferay Gradle Plugins].
 [LPS-69470]: https://issues.liferay.com/browse/LPS-69470
 [LPS-69488]: https://issues.liferay.com/browse/LPS-69488
 [LPS-69492]: https://issues.liferay.com/browse/LPS-69492
+[LPS-69501]: https://issues.liferay.com/browse/LPS-69501
 [LRDOCS-2594]: https://issues.liferay.com/browse/LRDOCS-2594
 [LRDOCS-2841]: https://issues.liferay.com/browse/LRDOCS-2841
 [LRDOCS-2981]: https://issues.liferay.com/browse/LRDOCS-2981